adds a new Overview component and moves titles to Heading component
