+++ src/docs/components/pages/Component.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import style from 'styled-components';
-import Example from './Example';
-import Props from './Props';
+import Example from '../Example';
+import Props from '../Props';
+import Heading from '../shared/Heading';
 
 const Wrapper = style.div`
     margin: 0px;
@@ -32,29 +33,6 @@ const Body = style.div`
     margin: 0px 88px;
 `;
 
-const Header = style.div`
-    display: block;
-    margin: 0px;
-    padding: 0px;
-    border-bottom: 1px solid #E6ECF1;
-    margin-bottom: 32px;
-`;
-
-const Title = style.div`
-    margin: 0px;
-    padding: 40px 0px;
-    flex: auto;
-    display: flex;
-`;
-
-const Text = style.h2`
-    display: block;
-    margin: 0px;
-    flex: 1 1 auto;
-    border-left: 2px solid #E6ECF1;
-    padding: 0 16px;
-    border-left-color: #2E5DF8;
-`;
 
 const Description = style.p`
     font-size: 14px;
@@ -72,7 +50,7 @@ const ExampleWrapper = style.div`
 `;
 
 
-const ComponentPage = ({ component }) => {
+const Component = ({ component }) => {
   const {
     name, description, props, examples,
   } = component;
@@ -80,11 +58,7 @@ const ComponentPage = ({ component }) => {
     <Wrapper>
       <Container>
         <Body>
-          <Header>
-            <Title>
-              <Text>{name}</Text>
-            </Title>
-          </Header>
+          <Heading text={name} />
           <Description>{description}</Description>
 
           <h3>
@@ -116,9 +90,9 @@ const ComponentPage = ({ component }) => {
   );
 };
 
-ComponentPage.propTypes = {
+Component.propTypes = {
   component: PropTypes.node.isRequired,
 };
 
 
-export default ComponentPage;
+export default Component;
