+++ src/docs/components/pages/Overview.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+export default class Overview extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+    };
+  }
+
+  render() {
+    return (
+      <div />
+    );
+  }
+}
+
+Overview.propTypes = {
+
+};
